requirements.txt
aiohttp~=3.11.10
alembic==1.13.0
alembic-postgresql-enum
anyio==4.9.0
apscheduler==3.10.4
asyncpg==0.30.0
Authlib
fastapi-utils
fastapi[all]~=0.112.2
greenlet>=2.0.0
httpx
jsonschema
pandas
pre-commit
psycopg2-binary~=2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pylint
pytest
pytest_mock
python-dotenv
python-multipart>=0.0.7
PyYAML
pyyaml~=6.0.2
redis
sqlalchemy~=2.0.41
starlette~=0.38.2
uvicorn
